[main] Funcional Test #1

+++ src/main/resources/SQL/post_sql.sql
@@ -2,7 +2,7 @@ CREATE TABLE post (
     `post_id` bigint NOT NULL AUTO_INCREMENT,
     `title` varchar(50) NOT NULL,
     `content` varchar(255) NOT NULL,
-    `user_id` bigint NOT NULL,
+    `user_id` int NOT NULL,
     `created_at` datetime NOT NULL,
     PRIMARY KEY (`post_id`)
 )
